int to size_t
changed an int to size_t, becasue a comparison between int and size_t was causing problems for emscripten

+++ Minigin/BenchMarkComponent.cpp
@@ -43,7 +43,7 @@ std::vector<float> dae::BenchmarkComponent::ProcessResults(const std::vector<flo
         std::vector<float> samples;
         for (int sampleIdx = 0; sampleIdx < numSamples; ++sampleIdx)
         {
-            int index = sampleIdx * numStepSizes + stepIdx;
+            size_t index = sampleIdx * numStepSizes + stepIdx;
             if (index < rawData.size())
             {
                 samples.push_back(rawData[index]);
